pyproject.toml
[project]
name = "clean-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "pysonar>=1.3.0.4086",
    "pyyaml>=6.0.3",
    "ruff>=0.14.14",
    "truststore>=0.10.4",
]
